Add individual product component and activate functionality
New "_IndividualProduct" component has been added to improve code reusability. Existing code in "DemoProduct" has been extracted to this new component. The "_EditDeleteBtn" component is updated to only show the buttons when it is active. This gives more control over when the buttons should be shown. A favourite count feature has been added to "DemoProduct" to keep track of the favourite products.