Add button click event in ParentComponent
A button has been added to the _ChildComponent.razor file and an event for on button click, 'OnButtonClick', has been added to the ParentComponent.razor which displays the message "Tangy Button clicked from Child Component" when the button is clicked. The event is linked from the child component to the parent through parameter passing.